Feat/73_productListing

+++ src/main/java/project/trendpick_pro/domain/product/controller/ProductController.java
@@ -3,9 +3,13 @@ package project.trendpick_pro.domain.product.controller;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import project.trendpick_pro.domain.common.base.rq.Rq;
+import project.trendpick_pro.domain.member.entity.Member;
+import project.trendpick_pro.domain.member.entity.RoleType;
 import project.trendpick_pro.domain.product.entity.dto.request.ProductSaveRequest;
 import project.trendpick_pro.domain.product.entity.dto.response.ProductResponse;
 import project.trendpick_pro.domain.product.service.ProductService;
@@ -18,10 +22,14 @@ import java.io.IOException;
 @RequestMapping("trendpick/products")
 public class ProductController {
 
-    private ProductService productService;
+    private final Rq rq;
+    private final ProductService productService;
 
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/register")
     public String registerProduct(@Valid ProductSaveRequest productSaveRequest, Model model) throws IOException {
+        if(!rq.getMember().getRole().equals(RoleType.BRAND_ADMIN))
+            throw new RuntimeException("상품 등록 권한이 없습니다.");
 
         ProductResponse productResponse = productService.register(productSaveRequest);
         model.addAttribute("productResponse", productResponse);
@@ -29,8 +37,11 @@ public class ProductController {
         return "redirect:/trendpick/products/" + productResponse.getId();
     }
 
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/edit/{productId}")
     public String modifyProduct(@PathVariable Long productId, @Valid ProductSaveRequest productSaveRequest, Model model) {
+        if(!rq.getMember().getRole().equals(RoleType.BRAND_ADMIN)) //추가로 해당 상품 브랜드 관리자인지도 체크
+            throw new RuntimeException("상품 수정 권한이 없습니다.");
 
         ProductResponse productResponse = productService.modify(productId, productSaveRequest);
         model.addAttribute("productResponse", productResponse);
@@ -38,8 +49,12 @@ public class ProductController {
         return "redirect:/trendpick/products/" + productResponse.getId();
     }
 
+    @PreAuthorize("isAuthenticated()")
     @DeleteMapping("/{productId}")
     public String deleteProduct(@PathVariable Long productId) {
+        if(!rq.getMember().getRole().equals(RoleType.BRAND_ADMIN)) //추가로 해당 상품 브랜드 관리자인지도 체크
+            throw new RuntimeException("상품 삭제 권한이 없습니다.");
+
         productService.delete(productId);
         return "redirect:/trendpick/products/list";
     }
@@ -51,9 +66,12 @@ public class ProductController {
     }
 
     @GetMapping("/list")
-    public String showAllProduct(@RequestParam("page") int offset, @RequestParam("main-category") String mainCategory,
-                                 @RequestParam("sub-category") String subCategory, Model model) {
-        model.addAttribute("productResponse", productService.showAll(offset, mainCategory, mainCategory));
+    public String showAllProduct(@RequestParam(value = "page", defaultValue = "0") int offset,
+                                 @RequestParam(value = "main-category") String mainCategory,
+                                 @RequestParam(value = "sub-category") String subCategory,
+                                 @RequestParam(value = "sort", defaultValue = "1") Integer sortCode,
+                                 Model model) {
+        model.addAttribute("productResponse", productService.showAll(offset, mainCategory, subCategory, sortCode));
         return "/trendpick/products/detailpage";
     }
 }
